Some more basic refactoring on writer calls for parallel passes
