+++ src/main/scala/com/github/plume/oss/passes/concurrent/PlumeContainsEdgePass.scala
@@ -2,17 +2,11 @@ package com.github.plume.oss.passes.concurrent
 
 import com.github.plume.oss.drivers.IDriver
 import com.github.plume.oss.passes.PlumeCpgPassBase
-import com.github.plume.oss.passes.concurrent.PlumeCfgCreationPass.producerQueueCapacity
-import com.github.plume.oss.passes.concurrent.PlumeConcurrentWriter.concurrentCreateApply
+import com.github.plume.oss.passes.concurrent.PlumeConcurrentCpgPass.concurrentCreateApply
 import io.shiftleft.codepropertygraph.Cpg
-import io.shiftleft.codepropertygraph.generated.nodes.{AstNode, Method}
+import io.shiftleft.codepropertygraph.generated.nodes.AstNode
 import io.shiftleft.passes.DiffGraph
 import io.shiftleft.semanticcpg.passes.base.ContainsEdgePass
-
-import scala.collection.mutable
-import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.duration.Duration
-import scala.concurrent.{Await, Future}
 
 object PlumeContainsEdgePass {
   private val producerQueueCapacity = 2 + 4 * Runtime.getRuntime.availableProcessors()
