+++ src/main/scala/com/github/plume/oss/passes/parallel/PlumeMethodStubCreator.scala
@@ -1,6 +1,7 @@
 package com.github.plume.oss.passes.parallel
 
 import com.github.plume.oss.drivers.IDriver
+import com.github.plume.oss.passes.parallel.PlumeParallelCpgPass.parallelWithWriter
 import com.github.plume.oss.passes.{IncrementalKeyPool, PlumeCpgPassBase}
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.passes.{DiffGraph, KeyPool, ParallelIteratorExecutor}
@@ -39,21 +40,8 @@ class PlumeMethodStubCreator(
     }
   }
 
-  private def withWriter[X](driver: IDriver)(f: PlumeParallelWriter => Unit): Unit = {
-    val writer       = new PlumeParallelWriter(driver, cpg)
-    val writerThread = new Thread(writer)
-    writerThread.setName("Writer")
-    writerThread.start()
-    try {
-      f(writer)
-    } catch {
-      case exception: Exception =>
-        baseLogger.warn("pass failed", exception)
-    } finally {
-      writer.enqueue(None, None)
-      writerThread.join()
-    }
-  }
+  private def withWriter[X](driver: IDriver)(f: PlumeParallelWriter => Unit): Unit =
+    parallelWithWriter(driver, f, cpg, baseLogger)
 
   private def enqueueInParallel(writer: PlumeParallelWriter): Unit = {
     withStartEndTimesLogged {
