+++ src/main/scala/com/github/plume/oss/passes/parallel/PlumeParallelCpgPass.scala
@@ -2,8 +2,10 @@ package com.github.plume.oss.passes.parallel
 
 import com.github.plume.oss.drivers.IDriver
 import com.github.plume.oss.passes.PlumeCpgPassBase
+import com.github.plume.oss.passes.parallel.PlumeParallelCpgPass.parallelWithWriter
 import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.passes.{KeyPool, ParallelCpgPass, ParallelIteratorExecutor}
+import org.slf4j.Logger
 
 abstract class PlumeParallelCpgPass[T](
     cpg: Cpg,
@@ -17,21 +19,8 @@ abstract class PlumeParallelCpgPass[T](
     }
   }
 
-  private def withWriter[X](driver: IDriver)(f: PlumeParallelWriter => Unit): Unit = {
-    val writer       = new PlumeParallelWriter(driver, cpg)
-    val writerThread = new Thread(writer)
-    writerThread.setName("Writer")
-    writerThread.start()
-    try {
-      f(writer)
-    } catch {
-      case exception: Exception =>
-        baseLogger.warn("pass failed", exception)
-    } finally {
-      writer.enqueue(None, None)
-      writerThread.join()
-    }
-  }
+  private def withWriter[X](driver: IDriver)(f: PlumeParallelWriter => Unit): Unit =
+    parallelWithWriter(driver, f, cpg, baseLogger)
 
   private def enqueueInParallel(writer: PlumeParallelWriter): Unit = {
     withStartEndTimesLogged {
@@ -73,4 +62,27 @@ abstract class PlumeParallelCpgPass[T](
     }
   }
 
+}
+
+object PlumeParallelCpgPass {
+  def parallelWithWriter[X](
+      driver: IDriver,
+      f: PlumeParallelWriter => Unit,
+      cpg: Cpg,
+      baseLogger: Logger
+  ): Unit = {
+    val writer       = new PlumeParallelWriter(driver, cpg)
+    val writerThread = new Thread(writer)
+    writerThread.setName("Writer")
+    writerThread.start()
+    try {
+      f(writer)
+    } catch {
+      case exception: Exception =>
+        baseLogger.warn("pass failed", exception)
+    } finally {
+      writer.enqueue(None, None)
+      writerThread.join()
+    }
+  }
 }
